Fix minor issues with integration tests.
- Add report files to list for cleanup on the power consumption test.
- Start scaling test at 2 nodes so All2All works properly.
- Add env var GEOPM_KEEP_FILES to optionally save all files generated
  during the test.

+++ test_integration/geopm_test_integration.py
@@ -53,11 +53,12 @@ class TestIntegration(unittest.TestCase):
 
     def tearDown(self):
         if sys.exc_info() == (None, None, None): # Will not be none if handling exception (i.e. failing test)
-            for ff in self._tmp_files:
-                try:
-                    os.remove(ff)
-                except OSError:
-                    pass
+            if os.getenv('GEOPM_KEEP_FILES') is None:
+                for ff in self._tmp_files:
+                    try:
+                        os.remove(ff)
+                    except OSError:
+                        pass
 
     def assertNear(self, a, b, epsilon=0.05):
         if abs(a - b) / a >= epsilon:
@@ -296,7 +297,7 @@ class TestIntegration(unittest.TestCase):
         """
         name = 'test_scaling'
         report_path = name + '.report'
-        num_node = 1
+        num_node = 2
         loop_count = 100
 
         app_conf = geopm_io.AppConf(name + '_app.config')
@@ -360,6 +361,9 @@ class TestIntegration(unittest.TestCase):
         launcher.set_num_rank(num_rank)
         launcher.write_log(name, 'Power cap = {}W'.format(self._options['power_budget']))
         launcher.run(name)
+
+        reports = [ff for ff in os.listdir('.') if fnmatch.fnmatch(ff, report_path + '*')]
+        self._tmp_files.extend(reports)
 
         traces = [ff for ff in os.listdir('.') if fnmatch.fnmatch(ff, trace_path + '*')]
         self._tmp_files.extend(traces)
